pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "research-agent"
version = "0.2.0"
description = "Template-constrained research agent with claim-graph contracts"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2",
]

[project.optional-dependencies]
retrieval = [
    "beautifulsoup4",
    "feedparser",
    "openai",
    "requests",
]
dev = [
    "pyyaml",
    "research-agent[retrieval]",
]

[project.scripts]
research-agent = "research_agent.cli.research:main"
claim-graph = "research_agent.cli.claim_graph:main"

[tool.setuptools.packages.find]
where = ["src"]
